Replace deployer registry with *Registry
